requirements.txt
Pillow==9.4.0
requests==2.28.2
aiogram==2.25.1
python-dotenv==0.21.1
python-telegram-bot-pagination==0.0.2
